feat(firmware): support 64-bit file headers for >4GB firmware
FileHeaderV3 now mirrors the real on-disk layout with split lo/hi u32
pairs for stored_length/original_length/offset, plus the previously
dropped offset_hi, unknown[64] and checksum fields. The per-file
offset/length accessors now return u64 (V3 combines lo|hi<<32, V1
widens its u32).

Widen the read path end-to-end: read_data_at_offset(offset: u64,
length: u64), and fix the truncation bug in
get_file_data_range_by_maintype_subtype where a u64 partition offset
was cast back to u32 right before seeking. Also widen
ImageInfo.image_size to u64, FileInfo fields to u64, and drop the now
redundant casts (incl. a clippy-driven ALIGNMENT_SIZE cleanup).

+++ src/tui/bridge.rs
@@ -91,7 +91,7 @@ pub fn load_firmware(path: &Path) -> Result<(OpenixPacker, u64, u32, Vec<String>
     let loaded =
         LoadedFirmware::load(path).map_err(|e| format!("Failed to load firmware: {}", e))?;
 
-    let size = loaded.image_info().image_size as u64;
+    let size = loaded.image_info().image_size;
     let num_files = loaded.image_info().num_files;
     let partition_names = loaded.partition_names().to_vec();
